refactor(storybook): resolve outstanding issues
- Resolve linting warnings and errors
- Refactor folder structure for components and stories (removed 'atoms')
- Resolve missing
