+++ src/stories/components/forms/Button.stories.ts
@@ -1,7 +1,7 @@
 // Button.stories.ts
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { Button } from '../../../components/atoms/Button';
+import { Button } from '../../../components/forms/Button';
 
 const meta: Meta<typeof Button> = {
   component: Button,
