+++ src/stories/components/forms/TextField.stories.ts
@@ -2,7 +2,7 @@
 
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { TextField } from '../../components/atoms/TextField';
+import { TextField } from '../../../components/forms/TextField';
 
 const meta: Meta<typeof TextField> = {
   component: TextField,
