+++ src/stories/components/forms/ToggleButton.stories.ts
@@ -2,7 +2,7 @@
 
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { ToggleButton } from '../../components/atoms/ToggleButton';
+import { ToggleButton } from '../../../components/forms/ToggleButton';
 
 const meta: Meta<typeof ToggleButton> = {
   component: ToggleButton,
